requirements.txt
numpy==1.16.4
lmdb==0.94
tensorboardX==1.2
tensorflow==1.15.2
tensorpack==0.9.4
tqdm==4.31.1
easydict==1.9
PyYAML==5.1.2
jsonlines==1.2.0
json-lines==0.5.0
matplotlib
Cython
python-prctl
msgpack
msgpack-numpy
opencv-python==4.2.0.34
botocore
pytorch-transformers==1.1.0
pycocotools
